+++ modules/commons/src/main/java/org/apache/ignite/internal/util/io/GridByteArrayOutputStream.java
@@ -22,8 +22,8 @@ import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.io.OutputStream;
 import java.util.Arrays;
+import org.apache.ignite.internal.util.CommonUtils;
 import org.apache.ignite.internal.util.typedef.internal.S;
-import org.apache.ignite.internal.util.typedef.internal.U;
 
 /**
  * This class defines output stream backed by byte array.
@@ -102,7 +102,7 @@ public class GridByteArrayOutputStream extends OutputStream {
         if (newCnt > buf.length)
             buf = Arrays.copyOf(buf, Math.max(buf.length << 1, newCnt));
 
-        U.arrayCopy(b, off, buf, cnt, len);
+        CommonUtils.arrayCopy(b, off, buf, cnt, len);
 
         cnt = newCnt;
     }
